fb - cd + usuwanie playlisty (poprawka)

+++ src/main/java/psk/isf/sts/service/playlist/PlaylistService.java
@@ -1,6 +1,7 @@
 package psk.isf.sts.service.playlist;
 
 import java.util.Collection;
+import java.util.Optional;
 
 import org.h2.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,19 +35,21 @@ public class PlaylistService {
 		return playlistRepo.save(playlist);
 	}
 
-	
-	public void deletePlaylist(User user, Long id) throws IllegalAccessException{
+	public void deletePlaylist(User user, Long id) throws IllegalAccessException {
 
-	Playlist playlist = playlistRepo.findById(id).get();
+		Optional<Playlist> playlistOptional = playlistRepo.findById(id);
+		if (!playlistOptional.isPresent())
+			return;
 
-    if(user.getId() != playlist.getUser().getId() ){ 
-      throw new IllegalAccessException("Nie masz uprawnien do usuniece tej playlisty");
-    }
-    
+		Playlist playlist = playlistOptional.get();
+
+		if (user.getId() != playlist.getUser().getId()) {
+			throw new IllegalAccessException("Nie masz uprawnien do usuniece tej playlisty");
+		}
+
 		playlistRepo.delete(playlist);
 	}
-	 
-	
+
 	public void validatePlaylist(PlaylistDTO dto) throws Exception {
 		if (StringUtils.isNullOrEmpty(dto.getName())) {
 			throw new Exception("Musisz podać nazwę playlisty");
